# Moises.botarro/creating cache for data scrubber