@DataDog/burrito 

This PR adds a cache on DataScrubber to avoid using the regex matching for already know processes.
After each process check, we add one cycle on the cache and reaching the `cacheTTL` limit, it's reset. 

This must decrease the number of calls to `regexp.(*machine).tryBacktrack` that came up on the top of function calls while profiling the process-agent.

The following benchmark compares the approach without and with cache for a host with 1200 processes when every process stay on the cache and there aren't new process coming up.

|benchmark           | old ns/op |     new ns/op |     delta|
|---|---|---|---|
|BenchmarkCache-2|     37082806|      555633        |-98.50%|

A test to check the cache use has been implemented too.